#20
Remove configurações do Nexus
Ajusta proteto para multi module
