+++ Exemplo-web/pom.xml
@@ -1,0 +1,138 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>br.com.alfa.posgraduacao</groupId>
+	<artifactId>exemplo</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<packaging>jar</packaging>
+
+	<name>exemplo</name>
+	<description>Demo project for Spring Boot</description>
+
+	<parent>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-parent</artifactId>
+		<version>1.5.8.RELEASE</version>
+		<relativePath/> <!-- lookup parent from repository -->
+	</parent>
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+		<java.version>1.8</java.version>
+	</properties>
+
+	<dependencies>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-data-jpa</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-devtools</artifactId>
+			
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
+		<dependency>
+		    <groupId>org.springframework.boot</groupId>
+		    <artifactId>spring-boot-starter-thymeleaf</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.postgresql</groupId>
+			<artifactId>postgresql</artifactId>
+			<scope>runtime</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.tiles</groupId>
+			<artifactId>tiles-jsp</artifactId>
+			<version>3.0.5</version>
+		</dependency>
+		<dependency>
+			<groupId>br.com.alfa.posgraduacao.exemplo</groupId>
+			<artifactId>core</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
+	</dependencies>
+
+	<build>
+		<plugins>
+		
+			<plugin>
+				<groupId>org.springframework.boot</groupId>
+				<artifactId>spring-boot-maven-plugin</artifactId>
+			 	
+			</plugin>
+			
+			 <plugin>
+		      <groupId>org.codehaus.mojo</groupId>
+		      <artifactId>exec-maven-plugin</artifactId>
+		      <executions>
+		        <execution>
+		          <id>exec-npm-install</id>
+		          <phase>generate-sources</phase>
+		          <configuration>
+		            <executable>npm.cmd</executable>
+		            <arguments>
+		              <argument>update</argument>
+		              
+		            </arguments>
+		            <workingDirectory>src/main/resources/static</workingDirectory>
+		                
+		          </configuration>
+		          <goals>
+		            <goal>exec</goal>
+		          </goals>
+		        </execution>
+		        <execution>
+		          <id>exec-bower-install</id>
+		          <phase>generate-sources</phase>
+		          <configuration>
+		            <executable>bower.cmd</executable>
+		            <arguments>
+		              <argument>update</argument>
+		            </arguments>
+		            <workingDirectory>src/main/resources/static</workingDirectory>
+		          </configuration>
+		          <goals>
+		            <goal>exec</goal>
+		          </goals>
+		        </execution>
+		       
+		      </executions>
+		      
+		    </plugin>
+		    <plugin>
+			    <artifactId>maven-clean-plugin</artifactId>
+			   
+			    <configuration>
+			      <filesets>
+			        <fileset>
+			          <directory>src/main/resources/static/components</directory>
+			          
+			        </fileset>
+			        <fileset>
+			          
+			          <directory>src/main/resources/static/node_modules</directory>
+			        </fileset>
+			      </filesets>
+			    </configuration>
+			  </plugin>
+				<plugin>
+				    <groupId>org.codehaus.sonar</groupId>
+				    <artifactId>sonar-maven-plugin</artifactId>
+				    <version>5.3</version>
+				</plugin>
+		</plugins>
+	</build>
+
+
+</project>
