+++ Exemplo-web/src/main/resources/static/.bowerrc
@@ -1,0 +1,3 @@
+{
+	"directory":"components"
+}
