+++ Exemplo-web/src/main/resources/static/.gitignore
@@ -1,0 +1,3 @@
+/components/
+
+/node_modules/
