+++ Exemplo-web/src/main/resources/static/exemplo/pessoa/pessoa.js
@@ -1,0 +1,12 @@
+/**
+ * 
+ */
+
+$(document).ready(function(){
+	  $('.exemplojs').exemplo();
+      $('#voltar').on('click',function(){
+    	 
+    	  window.location='/pessoa';
+      })
+   
+});
