--- src/main/resources/application.properties.padrao
@@ -1,5 +1,0 @@
-spring.datasource.url= jdbc:postgresql://localhost:5439/exemplo
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
