--- src/main/resources/static/.gitignore
@@ -1,3 +1,0 @@
-/components/
-
-/node_modules/
