--- src/main/resources/static/exemplo/pessoa/pessoa.js
@@ -1,12 +1,0 @@
-/**
- * 
- */
-
-$(document).ready(function(){
-	  $('.exemplojs').exemplo();
-      $('#voltar').on('click',function(){
-    	 
-    	  window.location='/pessoa';
-      })
-   
-});
