--- src/main/resources/templates/pessoa/pessoa.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-	<head th:include="thymeleaf/layout :: headerFragment">
-	<meta charset="ISO-8859-1"/>
-	<title>Insert title here</title>
-</head>
-<body>
-
-<div th:include="thymeleaf/layout :: menuFragment"></div>
-<div class="container">
-<form action="" th:object="${pessoa}" method="post">
-	<input type="hidden" name="_method" value="" th:attr="value=${method}"/>	
-		<div class="form-group">
-			<label for="codigo">Codigo
-			<input type="text" class="exemplojs form-control disabled" id="codigo" th:field="${pessoa.codigo}"  />
-			</label>
-		</div>
-	
-		<div class="form-group col-12">
-			<label>Nome
-				<input type="text" class="exemplojs form-control" th:field="${pessoa.nome}" />
-			</label>
-		</div>
-	
-		<div class="form-group">
-			<label>Idade
-				<input type="number" class="exemplojs form-control" th:field="${pessoa.idade}" />
-			</label>
-		</div>
-		<button type="submit" class="btn">Salvar</button>
-		<button type="button" class="btn" id="voltar">Voltar</button>
-</form>
-</div>
-<div th:include="thymeleaf/layout :: scriptFragment" th:remove="tag"></div>
-<script th:attr="src=@{/exemplo/pessoa/pessoa.js}" ></script>
-<script th:attr="src=@{/components/exemplo-js/exemplo.js}" ></script>
-</body>
-</html>
